feat(security): integrate SecurityModule and refactor guards
