+++ backend/src/auth/auth.controller.spec.ts
@@ -3,7 +3,7 @@ import { AuthController } from './auth.controller';
 import { AuthService } from './auth.service';
 import { UsersService } from '../users/users.service';
 import { AuthentificationInput } from './dto/auth-input-auth.dto';
-import { JwtAuthGuard } from './guards/jwt-auth.guard';
+import { JwtAuthGuard } from '../security/guards/jwt-auth.guard';
 import { ConfigService } from '@nestjs/config';
 
 describe('AuthController', () => {
