+++ backend/src/auth/auth.module.ts
@@ -2,27 +2,13 @@ import { Module } from '@nestjs/common';
 import { AuthService } from './auth.service';
 import { AuthController } from './auth.controller';
 import { UsersModule } from '../users/users.module';
-import { JwtModule } from '@nestjs/jwt';
-import { ConfigModule, ConfigService } from '@nestjs/config';
-import { forwardRef } from '@nestjs/common/utils/forward-ref.util';
+import { SecurityModule } from '../security/security.module';
+import { ConfigModule } from '@nestjs/config';
 
 @Module({
   controllers: [AuthController],
   providers: [AuthService],
-  imports: [
-    forwardRef(() => UsersModule),
-    ConfigModule,
-    JwtModule.registerAsync({
-      imports: [ConfigModule],
-      inject: [ConfigService],
-      useFactory: (config: ConfigService) => ({
-        secret: config.get<string>('JWT_SECRET'),
-        signOptions: {
-          expiresIn: config.get<string>('JWT_EXPIRES_IN') || '3d',
-        },
-      }),
-    }),
-  ],
-  exports: [AuthService, JwtModule],
+  imports: [UsersModule, SecurityModule, ConfigModule],
+  exports: [AuthService],
 })
 export class AuthModule {}
