+++ backend/src/security/guards/jwt-auth.guard.ts
@@ -1,4 +1,3 @@
-// auth/guards/jwt-auth.guard.ts
 import {
   CanActivate,
   ExecutionContext,
@@ -6,28 +5,28 @@ import {
   UnauthorizedException,
 } from '@nestjs/common';
 import { Request } from 'express';
-import { ConfigService } from '@nestjs/config';
-import { AuthService } from '../auth.service';
+import { JwtService } from '@nestjs/jwt';
+import { JwtPayload } from '../../auth/interfaces/jwt-payload.interface';
 
 @Injectable()
 export class JwtAuthGuard implements CanActivate {
-  constructor(
-    private readonly configService: ConfigService,
-    private authService: AuthService,
-  ) {}
+  constructor(private jwtService: JwtService) {}
 
-  canActivate(context: ExecutionContext): boolean {
+  async canActivate(context: ExecutionContext): Promise<boolean> {
     const request = context.switchToHttp().getRequest<Request>();
     const token = this.extractTokenFromHeader(request);
 
     if (!token) {
       throw new UnauthorizedException('Token not provided');
     }
-
-    const payload = this.authService.verifyToken(token);
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
-    (request as any).user = payload; // opcional: agrega el usuario a la request
-    console.log(payload);
+    try {
+      const payload: JwtPayload = await this.jwtService.verifyAsync(token);
+      // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
+      (request as any).user = payload; // opcional: agrega el usuario a la request
+      console.log(payload);
+    } catch {
+      throw new UnauthorizedException('Token not verified');
+    }
     return true;
   }
