+++ backend/src/security/security.module.ts
@@ -1,0 +1,30 @@
+import { Module } from '@nestjs/common';
+import { Reflector } from '@nestjs/core';
+import { JwtAuthGuard } from './guards/jwt-auth.guard';
+import { RoleGuard } from './guards/role.guard';
+import { JwtModule } from '@nestjs/jwt';
+import { ConfigModule, ConfigService } from '@nestjs/config';
+
+@Module({
+  imports: [
+    JwtModule.registerAsync({
+      imports: [ConfigModule],
+      inject: [ConfigService],
+      useFactory: (config: ConfigService) => {
+        const secret = config.get<string>('JWT_SECRET');
+        if (!secret) {
+          throw new Error('JWT_SECRET is not defined in environment variables');
+        }
+        return {
+          secret,
+          signOptions: {
+            expiresIn: config.get<string>('JWT_EXPIRES_IN') || '3d',
+          },
+        };
+      },
+    }),
+  ],
+  providers: [JwtAuthGuard, RoleGuard, Reflector],
+  exports: [JwtAuthGuard, RoleGuard, JwtModule],
+})
+export class SecurityModule {}
