+++ backend/src/users/users.module.ts
@@ -3,11 +3,10 @@ import { UsersService } from './users.service';
 import { UsersController } from './users.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { User } from './entities/user.entity';
-import { AuthModule } from '../auth/auth.module';
-import { CommonModule } from '../common/common.module';
+import { SecurityModule } from '../security/security.module';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([User]), CommonModule, AuthModule],
+  imports: [TypeOrmModule.forFeature([User]), SecurityModule],
   controllers: [UsersController],
   providers: [UsersService],
   exports: [UsersService],
